avoid duplicated code
